do some refactors
