+++ src/pitch_sequencing/ml/data/count_state.py
@@ -8,21 +8,7 @@ from torch.utils.data import Dataset
 
 from pitch_sequencing.ml.tokenizers.pitch_sequence import PitchSequenceWithCountTokenizer, SeparateSequenceTokenizer
 from pitch_sequencing.ml.data.last_pitch import extract_last_element_from_csv_seq
-
-@dataclass
-class SingularSequence:
-    """
-    src: tokenized sequence of type long. Size 1xN
-    src_mask: boolean tensor indicating padding of src. Size 1xN.
-    """
-    src: torch.Tensor
-    src_mask: torch.Tensor
-
-    def to(self, device: torch.device) -> 'SingularSequence':
-        self.src = self.src.to(device)
-        self.src_mask = self.src_mask.to(device)
-
-        return self
+from pitch_sequencing.ml.data.sequences import SingularSequence
 
 @dataclass
 class PitchCountSequences:
@@ -44,15 +30,6 @@ class PitchCountSequences:
 
 
 # TODO(kaelen) figure out how to not write these for each type
-def collate_interleaved_and_target(batch) -> typing.Tuple[SingularSequence, torch.Tensor]:
-    seq_data_list = [item[0] for item in batch]
-    targets = torch.stack([item[1] for item in batch])
-
-    srcs = torch.stack([data.src for data in seq_data_list])
-    src_masks = torch.stack([data.src_mask for data in seq_data_list])
-
-    return SingularSequence(srcs, src_masks), targets
-
 def collate_pitch_count_seqs_and_target(batch) -> typing.Tuple[PitchCountSequences, torch.Tensor]:
     seq_data_list = [item[0] for item in batch]
     targets = torch.stack([item[1] for item in batch])
